Move permissions interfaces to sdk-model
- move from sdk-backend-spi to sdk-model, because it's part of the domain model

JIRA: RAIL-4109
