+++ libs/sdk-backend-spi/api/sdk-backend-spi.api.md
@@ -57,6 +57,7 @@ import { IWidgetAlert as IWidgetAlert_2 } from '@gooddata/sdk-model';
 import { IWidgetAlertDefinition as IWidgetAlertDefinition_2 } from '@gooddata/sdk-model';
 import { IWidgetDefinition as IWidgetDefinition_2 } from '@gooddata/sdk-model';
 import { IWidgetReferences as IWidgetReferences_2 } from '@gooddata/sdk-model';
+import { IWorkspacePermissions as IWorkspacePermissions_2 } from '@gooddata/sdk-model';
 import { IWorkspaceUser as IWorkspaceUser_2 } from '@gooddata/sdk-model';
 import { IWorkspaceUserGroup as IWorkspaceUserGroup_2 } from '@gooddata/sdk-model';
 import * as m from '@gooddata/sdk-model';
@@ -1737,14 +1738,12 @@ export interface IWorkspaceMeasuresService {
     updateMeasure(measure: IMeasureMetadataObject_2): Promise<IMeasureMetadataObject_2>;
 }
 
-// @public
-export type IWorkspacePermissions = {
-    [permission in WorkspacePermission]: boolean;
-};
+// @public @deprecated
+export type IWorkspacePermissions = m.IWorkspacePermissions;
 
 // @public
 export interface IWorkspacePermissionsService {
-    getPermissionsForCurrentUser(): Promise<IWorkspacePermissions>;
+    getPermissionsForCurrentUser(): Promise<IWorkspacePermissions_2>;
 }
 
 // @public
@@ -1979,83 +1978,7 @@ export const widgetType: typeof m.widgetType;
 // @public @deprecated
 export const widgetUri: typeof m.widgetUri;
 
-// @public
-export type WorkspacePermission =
-/**
-* Whether the current user has permissions to run MAQL DDL and DML, access a workspace staging directory.
-*/
-"canInitData"
-/**
-* Whether the current user has permissions to upload CSV files via CSV Uploader.
-*/
-| "canUploadNonProductionCSV"
-/**
-* Whether the current user has permissions to download a complete report.
-*/
-| "canExecuteRaw"
-/**
-* Whether the current user has permissions to export the report.
-*/
-| "canExportReport"
-/**
-* Whether the current user has permissions to access GoodData portal directly (user can log in).
-*/
-| "canAccessWorkbench"
-/**
-* Whether the current user has permissions to create a report object via API.
-*/
-| "canCreateReport"
-/**
-* Whether the current user has permissions to create a KPI object, KPI widget object, and an insight object via API.
-*/
-| "canCreateVisualization"
-/**
-* Whether the current user has permissions to create a KPI dashboard object via API.
-*/
-| "canCreateAnalyticalDashboard"
-/**
-* Whether the current user has permissions to modify and delete a metric, run MAQL DDL, run the MAQL validator, change metric visibility via the `unlisted` flag.
-*/
-| "canManageMetric"
-/**
-* Whether the current user has permissions to modify and delete a report object, change report visibility via the `unlisted` flag.
-*/
-| "canManageReport"
-/**
-* Whether the current user has permissions to modify and delete a KPI dashboard object.
-*/
-| "canManageAnalyticalDashboard"
-/**
-* Whether the current user has permissions to modify workspace metadata, see the workspace token, lock and unlock objects, delete locked objects, set and unset the restricted flag on objects, clear cache, delete a workspace.
-*/
-| "canManageProject"
-/**
-* Whether the current user has permissions to create a scheduled email object and a KPI alert object.
-*/
-| "canCreateScheduledMail"
-/**
-* Whether the current user has permissions to list users, roles, and permissions.
-*/
-| "canListUsersInProject"
-/**
-* Whether the current user has permissions to modify and delete a domain, run MAQL DDL.
-*/
-| "canManageDomain"
-/**
-* Whether the current user has permissions to invite a user to a workspace or delete an invitation.
-*/
-| "canInviteUserToProject"
-/**
-* Whether the current user has permissions to run uploads, load date dimensions, access a workspace staging directory.
-*/
-| "canRefreshData"
-/**
-* Whether the current user has permissions to add, remove, and list ACLs (Access Control Lists) on an object.
-*/
-| "canManageACL"
-/**
-* Whether the current user has permissions to manage scheduled email objects.
-*/
-| "canManageScheduledMail";
+// @public @deprecated
+export type WorkspacePermission = m.WorkspacePermission;
 
 ```
