+++ libs/sdk-backend-spi/src/deprecated/permissions.ts
@@ -1,0 +1,19 @@
+// (C) 2019-2022 GoodData Corporation
+/* eslint-disable @typescript-eslint/no-empty-interface, import/namespace */
+import * as m from "@gooddata/sdk-model";
+
+/**
+ * Workspace permission types
+ *
+ * @deprecated Use {@link @gooddata/sdk-model#WorkspacePermission}
+ * @public
+ */
+export type WorkspacePermission = m.WorkspacePermission;
+
+/**
+ * Dictionary of workspace permissions
+ *
+ * @deprecated Use {@link @gooddata/sdk-model#IWorkspacePermissions}
+ * @public
+ */
+export type IWorkspacePermissions = m.IWorkspacePermissions;
